exercises 4.15-4.23

+++ osa4/controllers/blogs.js
@@ -1,22 +1,31 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
     const body = request.body
+
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.json(savedBlog.toJSON())
 })
 
@@ -29,9 +38,16 @@ blogsRouter.get('/:id', async (request, response, next) => {
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if(!blog) {
+        return response.status(404).end()
+    } else if (user.id.toString() === blog.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
@@ -48,7 +64,6 @@ blogsRouter.put('/:id', async (request, response, next) => {
         .then(updatedBlog => {
             response.json(updatedBlog.toJSON())
         })
-        .catch(error => next(error))
 })
 
 module.exports = blogsRouter
